tui: Let ignore-file actions choose the destination file
The terminal file browser lets users mark files for actions. A block action
tells Git Stage Batch to ignore a file, so the browser also asks whether it
belongs in the repository-local exclude file or a tracked ignore file and
then changes that ignore state itself.

The browser should coordinate file selection and marked actions. Choosing
ignore storage and changing it are block-command responsibilities already
needed by the single-file block path.

This commit begins separating navigation from ignore-file changes by moving
the destination prompt into the block-action module and making the browser
delegate each marked block request to that module.

The next commit will update interaction and architecture tests to enforce
that delegation.
